server/requirements.txt
fastapi==0.115.0
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.31.0
